+++ src/index.ts
@@ -127,6 +127,7 @@ class Vditor extends VditorMethod {
                     mergedOptions.after();
                 }
             });
+        addScript(`${mergedOptions.cdn}/dist/js/icons/${mergedOptions.icon}.js`, "vditorIconScript");
     }
 
     /** 设置主题 */